Minor changes to Prefix scan. Removed use of separate offset buffer, doing prefix scan on the same buffer now. Added back face culling.

+++ ParticlePhysics/GridFrag.glsl
@@ -1,5 +1,5 @@
-in highp vec4 col;
-out highp vec4 outputColor;
+in lowp vec4 col;
+out lowp vec4 outputColor;
 
 void main()
 {
