+++ ParticlePhysics/SolidFrag.glsl
@@ -1,12 +1,12 @@
-in highp vec4 col;
+in lowp vec4 col;
 in highp vec3 normal;
 
 uniform highp float fillShader;
-out highp vec4 outputColor;
+out lowp vec4 outputColor;
 
 void main()
 {
-  highp vec4 scaledColor = col + vec4(0.4f, 0.4f, 0.4f, 0.0f);
+  lowp vec4 scaledColor = col + vec4(0.4f, 0.4f, 0.4f, 0.0f);
   if (fillShader == 1.f)
   {
     outputColor = scaledColor * clamp(dot(normal, vec3(0.f, 1.f, 0.f)), 0.2f, 1.f);
